server/requirements.txt
pytest
requests
python-firebase
